requirements.txt
fastapi
uvicorn
pydantic
pandas
requests
openpyxl
python-dateutil
pytest
httpx